Refactor person package to family package
