+++ src/main/java/seedu/planitarium/PlanITarium.java
@@ -5,9 +5,8 @@ import java.util.Scanner;
 
 import seedu.planitarium.commands.Command;
 import seedu.planitarium.commands.CommandFactory;
-import seedu.planitarium.exceptions.PlanITariumException;
 import seedu.planitarium.global.Help;
-import seedu.planitarium.person.Family;
+import seedu.planitarium.family.Family;
 import seedu.planitarium.global.UI;
 import seedu.planitarium.storage.Storage;
