+++ src/main/java/seedu/planitarium/commands/ExitCommand.java
@@ -2,7 +2,7 @@ package seedu.planitarium.commands;
 
 import seedu.planitarium.ProjectLogger;
 import seedu.planitarium.global.UI;
-import seedu.planitarium.person.Family;
+import seedu.planitarium.family.Family;
 import seedu.planitarium.storage.Storage;
 
 import java.util.logging.Level;
