+++ src/main/java/seedu/planitarium/commands/HelpCommand.java
@@ -2,7 +2,7 @@ package seedu.planitarium.commands;
 
 import seedu.planitarium.ProjectLogger;
 import seedu.planitarium.global.UI;
-import seedu.planitarium.person.Family;
+import seedu.planitarium.family.Family;
 
 import java.util.logging.Level;
